LPS-106617 Moving the reducer
